third_party/blink/web_tests/fast/xpath/evaluate-twice-expected.txt

PASS expr.evaluate(doc.documentElement.firstChild, XPathResult.STRING_TYPE, null).stringValue is '1'
PASS expr.evaluate(doc.documentElement.firstChild.nextSibling, XPathResult.STRING_TYPE, null).stringValue is '1'
PASS expr.evaluate(doc.documentElement.firstChild, XPathResult.STRING_TYPE, null).stringValue is '1'
PASS expr.evaluate(doc.documentElement.firstChild.nextSibling, XPathResult.STRING_TYPE, null).stringValue is '2'
PASS expr.evaluate(doc.documentElement.firstChild, XPathResult.STRING_TYPE, null).stringValue is '1'
PASS expr.evaluate(doc.documentElement.firstChild.nextSibling, XPathResult.STRING_TYPE, null).stringValue is '2'
PASS expr.evaluate(doc.documentElement.firstChild, XPathResult.NUMBER_TYPE, null).numberValue is 2
PASS expr.evaluate(doc.documentElement.firstChild.nextSibling, XPathResult.NUMBER_TYPE, null).numberValue is 3
PASS successfullyParsed is true

TEST COMPLETE
